updated texture params
